// libs/core/error/common/src/lib/translate-validation-errors.enum.ts
export const enum TranslateValidationErrorsEnum {
    EmailAlreadyTaken = 'email_alredy_taken',
    GoalIdAlreadyTaken = 'goal_id_already_taken',
    BlankEmail = 'blank_email',
    IncorrectEmail = 'incorrect_email',
    BlankPassword = 'blank_password',
    NoUser = 'no_user',
    ValidEmail = 'valid_email',
    AccountIsPendingApproval = 'account_is_pending_approval',
    UnavailableLogin = 'unavailable_login',
    PaymentMethodBlank = 'payment_method_blank',
    AmountShouldHigherPaymentThreshold = 'amount_should_higher_payment_threshold'
}

export const enum ValidationErrorsTranslateEnum {
    EmailAlreadyTaken = 'server_validation.email_alredy_taken',
    GoalIdAlreadyTaken = 'server_validation.goal_id_already_taken',
    BlankEmail = 'server_validation.blank_email',
    IncorrectEmail = 'server_validation.incorrect_email',
    BlankPassword = 'server_validation.blank_password',
    NoUser = 'server_validation.no_user',
    ValidEmail = 'server_validation.valid_email',
    AccountIsPendingApproval = 'server_validation.account_is_pending_approval',
    UnavailableLogin = 'server_validation.unavailable_login',
    PaymentMethodBlank = 'server_validation.payment_method_blank',
    AmountShouldHigherPaymentThreshold = 'server_validation.amount_should_higher_payment_threshold'
}